bot.py
import os
import discord

from discord.ext import commands
from dotenv import load_dotenv

load_dotenv()
TOKEN = os.getenv('DISCORD_TOKEN')
GUILD = os.getenv('DISCORD_GUILD')

startup_ext = ["trivia"]

bot = commands.Bot(command_prefix='!')

@bot.event
async def on_ready():
  guild = discord.utils.find(lambda g: g.name == GUILD, bot.guilds)
  print(
      f'{bot.user} is connected to: \n'
      f'{guild.name}(id: {guild.id})'
  )       

if __name__ == "__main__":
    for ext in startup_ext:
        try:
            bot.load_extension(ext)
        except Exception as e:
            exc = '{}: {}'.format(type(e).__name__, e)
            print('Failed to load extension {}\n{}'.format(ext, exc))


bot.run(TOKEN)